+++ src/Visualizer/Grid/Grid.jsx
@@ -4,31 +4,11 @@ import "./Grid.css";
 export default class Grid extends Component {
   constructor(props) {
     super(props);
-    this.state = {
-      grid: []
-    };
-  }
-
-  componentDidMount() {
-    const grid = this.getInitialGrid();
-    this.setState({ grid });
-  }
-
-  getInitialGrid() {
-    const grid = [];
-
-    for (let row = 0; row < 17; row++) {
-      const currentRow = [];
-      for (let col = 0; col < 70; col++) {
-        currentRow.push(0);
-      }
-      grid.push(currentRow);
-    }
-    return grid;
+    this.state = {};
   }
 
   render() {
-    const grid = this.state.grid;
+    const grid = this.props.grid;
     return (
       <>
         <table id="grid" onMouseLeave={this.props.setMouseDownFalse}>