+++ src/Visualizer/Menu/Menu.jsx
@@ -24,12 +24,19 @@ export default class Menu extends Component {
           onClick={() => this.props.selectNode(this.props.nodes.END_NODE)}>
           End
         </button>
-        <button type="button" onClick={() => this.props.runDijkstra()}>
+        <button type="button" onClick={() => this.props.runAlgo()}>
           Run
         </button>
         <button type="button" onClick={() => this.props.clearBoard()}>
           Clear Board
         </button>
+        <select
+          id="algos"
+          value={undefined}
+          onChange={event => this.props.updateAlgo(event.target.value)}>
+          <option hidden>Choose an Algorithm</option>
+          <option value="Dijkstra">Dijkstra</option>
+        </select>
       </div>
     );
   }