+++ src/Visualizer/Grid/Grid.jsx
@@ -1,45 +1,40 @@
-import React, { Component } from "react";
+import React from "react";
 import "./Grid.css";
 
-export default class Grid extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
+const Grid = props => {
+  const grid = props.grid;
+  return (
+    <>
+      <table id="grid" onMouseLeave={props.setMouseDownFalse}>
+        <tbody id="tableBody">
+          {grid.map((row, rowIdx) => {
+            return (
+              <tr key={rowIdx} id={`row_${rowIdx}`}>
+                {row.map((cell, cellIdx) => {
+                  return (
+                    <td
+                      key={cellIdx}
+                      id={`node_${rowIdx}_${cellIdx}`}
+                      className="unvisited"
+                      onMouseDown={event => {
+                        props.mouseEventHandler(event);
+                        props.setMouseDownTrue();
+                      }}
+                      onMouseUp={props.setMouseDownFalse}
+                      onMouseOver={event => {
+                        if (props.mouseDown) {
+                          props.mouseEventHandler(event);
+                        }
+                      }}></td>
+                  );
+                })}
+              </tr>
+            );
+          })}
+        </tbody>
+      </table>
+    </>
+  );
+};
 
-  render() {
-    const grid = this.props.grid;
-    return (
-      <>
-        <table id="grid" onMouseLeave={this.props.setMouseDownFalse}>
-          <tbody id="tableBody">
-            {grid.map((row, rowIdx) => {
-              return (
-                <tr key={rowIdx} id={`row_${rowIdx}`}>
-                  {row.map((cell, cellIdx) => {
-                    return (
-                      <td
-                        key={cellIdx}
-                        id={`node_${rowIdx}_${cellIdx}`}
-                        className="unvisited"
-                        onMouseDown={event => {
-                          this.props.mouseEventHandler(event);
-                          this.props.setMouseDownTrue();
-                        }}
-                        onMouseUp={this.props.setMouseDownFalse}
-                        onMouseOver={event => {
-                          if (this.props.mouseDown) {
-                            this.props.mouseEventHandler(event);
-                          }
-                        }}></td>
-                    );
-                  })}
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-      </>
-    );
-  }
-}
+export default Grid;