+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 - Initial grid is now displayed
 - There is a button for each available node
-- There is visual feedback when the a grid cell is clicked (the feedback is different for each node selected)
+- There is visual feedback when a grid cell is clicked (the feedback is different for each node selected)
 - Limitation to the number of start and end nodes (there can only be 1 of each)
 - A node can't be placed over an existing node (className != unvisited)
+- It is possible to add walls by clicking a cell and then dragging the mouse over the grid